move jet-specific code into jet crate
