+++ jet/cli/src/cli.rs
@@ -1,9 +1,13 @@
 use std::path::PathBuf;
 
+use custody_anchor_lang::AnchorDeserialize;
 use ::jet::state::MarketFlags;
 use anchor_client::solana_sdk::pubkey::Pubkey;
 use anyhow::Result;
 use clap::Parser;
+use multisig_anchor_client::Program;
+use multisig_anchor_spl::token::Mint;
+use multisig_anchor_spl::token::TokenAccount;
 use multisig_client::cli::run_bpf_proposal;
 use multisig_client::cli::run_multisig_command;
 use multisig_client::cli::run_multisig_proposal;
@@ -11,13 +15,15 @@ use multisig_client::cli::run_token_proposal;
 use multisig_client::cli::MISSING_MULTISIG;
 use multisig_client::config::load;
 use multisig_client::nested_subcommands;
+use multisig_client::service::ProposalInspector;
 use multisig_client::{
     cli::{BpfProposal, MultisigCommand, MultisigProposal, TokenAction, TokenProposal},
     service::MultisigService,
 };
 use paste::paste;
+use serum_multisig::Transaction;
 
-use crate::propose::custody;
+use crate::propose::custody as custody_proposals;
 use crate::propose::jet;
 use crate::propose::jet::ReserveParameters;
 
@@ -153,15 +159,85 @@ pub fn run_custody_proposal(
     match job {
         CustodyProposal::GenerateTokenMint(cmd) => {
             let key =
-                custody::propose_custody_generate_token_mint(&service, multisig, cmd.mint_key)?;
+                custody_proposals::propose_custody_generate_token_mint(&service, multisig, cmd.mint_key)?;
             println!("{}", key);
         }
         CustodyProposal::TransferTokens(cmd) => {
-            let key = custody::propose_custody_transfer_tokens(
+            let key = custody_proposals::propose_custody_transfer_tokens(
                 &service, multisig, cmd.source, cmd.target, cmd.amount,
             )?;
             println!("{}", key);
         }
     }
     Ok(())
+}
+
+pub struct JetProposalInspector;
+
+impl ProposalInspector for JetProposalInspector {
+    fn inspect_proposal(&self, program: &Program, proposed_tx: &Transaction) -> Result<bool> {
+        match proposed_tx.program_id {
+            pid if pid == custody::ID => {
+                let mut instr_hash = [0u8; 8];
+                instr_hash.copy_from_slice(&proposed_tx.data[..8]);
+
+                match instr_hash {
+                    hash if hash == instr_sighash("generate_token_mint") => {
+                        println!("Proposal to generate initial tokens");
+
+                        let mint = proposed_tx.accounts[0].pubkey;
+                        println!("Proposed mint: {}", mint);
+
+                        return Ok(true);
+                    }
+
+                    hash if hash == instr_sighash("transfer_funds") => {
+                        println!("Proposal to transfer funds to an account");
+
+                        let args = custody::instruction::TransferFunds::try_from_slice(
+                            &proposed_tx.data[8..],
+                        )?;
+                        let vault = proposed_tx.accounts[0].pubkey;
+                        let target = proposed_tx.accounts[1].pubkey;
+
+                        let vault_account = program.account::<TokenAccount>(vault)?;
+                        let mint_account =
+                            program.account::<Mint>(vault_account.mint)?;
+
+                        let base = 10f64.powf(mint_account.decimals as f64);
+                        let proposed_amount = (args.amount as f64) / base;
+                        let vault_amount = (vault_account.amount as f64) / base;
+
+                        println!("Transferring from: {}", vault);
+                        println!("Custodied amount: {}", vault_amount);
+                        println!(
+                            "Custodied remaining after the transfer: {}",
+                            vault_amount - proposed_amount
+                        );
+                        println!();
+                        println!("**TRANSFER AMOUNT**: {}", proposed_amount);
+                        println!();
+                        println!("**TRANSFER TO**: {}", target);
+
+                        Ok(true)
+                    }
+
+                    _ => Ok(false),
+                }
+            }
+
+            _ => Ok(false),
+        }
+    }
+}
+
+fn instr_sighash(name: &str) -> [u8; 8] {
+    let preimage = format!("global:{}", name);
+
+    let mut result = [0u8; 8];
+
+    result.copy_from_slice(
+        &anchor_client::solana_sdk::hash::hash(preimage.as_bytes()).to_bytes()[..8],
+    );
+    result
 }
